# chore(ci): link with mold in build-tool