--- a/.github/actions/build_debug/action.yml
+++ b/.github/actions/build_debug/action.yml
@@ -27,7 +27,7 @@ runs:
       if: "contains(inputs.target, 'linux')"
       uses: ./.github/actions/setup_build_tool
       with:
-        image: datafuselabs/build-tool:multiarch
+        image: everpcpc/build-tool:multiarch
 
     # If you need to reset the cache version, increment the number after `v`
     - uses: Swatinem/rust-cache@v1
@@ -36,6 +36,10 @@ runs:
 
     - shell: bash
       run: cargo build --target ${{ inputs.target }}
+
+    - shell: bash
+      if: "contains(inputs.target, 'linux')"
+      run: readelf -p .comment ./target/${{ inputs.target }}/debug/databend-query
 
     - name: Upload artifact
       uses: actions/upload-artifact@v2