--- a/.github/actions/build_debug/action.yml
+++ b/.github/actions/build_debug/action.yml
@@ -27,7 +27,7 @@ runs:
       if: "contains(inputs.target, 'linux')"
       uses: ./.github/actions/setup_build_tool
       with:
-        image: everpcpc/build-tool:multiarch
+        image: datafuselabs/build-tool:multiarch
 
     # If you need to reset the cache version, increment the number after `v`
     - uses: Swatinem/rust-cache@v1