I hereby agree to the terms of the CLA available at: https://databend.rs/dev/policies/cla/

## Summary

link with mold for debug linux gnu


### with mold

on x86_64-unknown-linux-gnu
```
 Finished dev [unoptimized + debuginfo] target(s) in 3m 26s

String dump of section '.comment':
  [     0]  GCC: (Debian 10.2.1-6) 10.2.1 20210110
  [    27]  mold 1.1.1 (c1af9c97e763a1d521889191a83b27bd6c43b351; compatible with GNU ld)
  [    75]  clang version 13.0.0
```

on aarch64-unknown-linux-gnu
```
Finished dev [unoptimized + debuginfo] target(s) in 3m 38s

String dump of section '.comment':
  [     0]  GCC: (Debian 10.2.1-6) 10.2.1 20210110
  [    27]  mold 1.1.1 (c1af9c97e763a1d521889191a83b27bd6c43b351; compatible with GNU ld)
  [    75]  GCC: (crosstool-NG ) 5.2.0
```

### without mold

on x86_64-unknown-linux-gnu
```
Finished dev [unoptimized + debuginfo] target(s) in 5m 53s

String dump of section '.comment':
  [     0]  GCC: (Debian 10.2.1-6) 10.2.1 20210110
  [    27]  clang version 13.0.0
```

on aarch64-unknown-linux-gnu
```
Finished dev [unoptimized + debuginfo] target(s) in 5m 11s

String dump of section '.comment':
  [     0]  GCC: (Debian 10.2.1-6) 10.2.1 20210110
  [    27]  GCC: (crosstool-NG ) 5.2.0
```



## Changelog

- Build/Testing/CI

## Related Issues

Fixes #4180

## Test Plan

Unit Tests

Stateless Tests