@Xuanwo 2022-04-03
@ClSlaid used to work on the same issue https://github.com/datafuselabs/databend/pull/4196, invited for a review, and thanks for his hard work!