@ClSlaid 2022-04-03
The PR seems to work by replacing the GNU `Gold` with `mold`. It's good but, will it also replace the builder for `aarch64-linux-unknown-gnu`? @Xuanwo