@everpcpc 2022-04-03
> The PR seems to work by replacing the GNU `Gold` with `mold`. It's good but, will it also replace the builder for `aarch64-linux-unknown-gnu`? @Xuanwo

Yes, it will replace ld under `x86-64` and `aarch64`.
They both working with the mold binary release:
```
mold: supported targets: elf32-i386 elf64-x86-64 elf64-littleaarch64
mold: supported emulations: elf_i386 elf_x86_64 aarch64linux aarch64elf
```